% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{authenticate
\code{authenticate} should be used in context of}
\usage{
authenticate(req, res, is_authed_fun, ...)
}
\arguments{
\item{req}{plumber request object}

\item{res}{plumber response object}

\item{is_authed_fun}{function. Function to check whether API call is authenticated / authorized.
Use any of sealr's is_authed_* functions or your own custom function. See Details for requirements for custom functions.}

\item{...}{arguments to be passed down to the is_authed_fun function.}
}
\description{
authenticate
\code{authenticate} should be used in context of
}
\details{
Custom is_authed_fun functions should return a list with the following elements:
\itemize{
  \item is_authed: TRUE or FALSE. Result of the check of is_authed_fun.
  \item status: character. short description of HTTP status code
  \item code: integer. HTTP status code
  \item message: character. Longer description.
}

You can use the helper functions \code{\link{is_authed_return_list}} and \code{\link{is_authed_return_list_401}}
to generate those lists in your custom function.
}
