% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_vars.R
\name{catch_vars}
\alias{catch_vars}
\alias{catch_varnames}
\alias{catch_varname}
\title{Catch-up data frame variable names and value}
\usage{
catch_varnames(data)

catch_varname(data, variable)
}
\arguments{
\item{data}{A data frame.}

\item{variable}{target data frame's column name (\code{character})}
}
\description{
Catch-up data frame variable names and value
}
\examples{
\dontrun{
catch_varnames(mtcars)

# Evaluate
library(glue)
evaluate(catch_varname(mtcars, "mpg"), .GlobalEnv)
}
}
