% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_jwt.R
\name{check_all_claims}
\alias{check_all_claims}
\title{This function checks that all claims passed in the \code{claims} argument of
the \code{\link{is_authed_jwt}} function are correct.}
\usage{
check_all_claims(payload, claims)
}
\arguments{
\item{payload}{JWT payload extracted with jose::jwt_decode_hmac.}

\item{claims}{named list of claims to check in the JWT. Claims can be nested.}
}
\value{
TRUE if the all claims are present in the JWT, FALSE if not.
}
\description{
This function checks that all claims passed in the \code{claims} argument of
the \code{\link{is_authed_jwt}} function are correct.
}
