% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_bearer_token}
\alias{clean_bearer_token}
\title{clean_bearer_token}
\usage{
clean_bearer_token(token)
}
\arguments{
\item{token}{character. The token to be handled.}
}
\value{
the cleaned token.
}
\description{
helper function that removes 'Bearer' from a bearer scheme token and trims whitespace.
}
