% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound.R
\name{compound}
\alias{compound}
\title{Storage R object to a temporary environment}
\description{
Storage R object to a temporary environment
}
\examples{
\dontrun{
compound(letters)
compound(iris)

x <- c(1, 3, 5)
x \%>\% compound()
}
}
