% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doggo.R
\name{doggo}
\alias{doggo}
\title{A Dog Function}
\usage{
doggo(num)
}
\arguments{
\item{num}{Do you love dogs? The number corresponds to how many dog pictures you would like to see to cheer you up.}
}
\description{
This function allows you to express your love of dogs. This is a wrapper for the pupR package from https://github.com/melissanjohnson/pupR
}
\examples{
doggo(10) #this gives you 10 dog pictures from pupR
}
\keyword{dogs}
