% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadSentinel2L1C.R
\name{DownloadSentinel2L1C}
\alias{DownloadSentinel2L1C}
\title{Download Sentinel 2 L1C dataset from google cloud.}
\usage{
DownloadSentinel2L1C(tilename, folder = folder)
}
\arguments{
\item{tilename}{what tile do you want to download}
}
\description{
This function requires Google Cloud Storage SDK, you can download and install this SDK
from google cloud website.
Use this carefully, cause once it started, it's not easy to stop.
}
\examples{
DownloadSentinel2L1C("17SNA")
}
\keyword{sentinel}
