% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_token_from_req}
\alias{get_token_from_req}
\title{get_token_from_req}
\usage{
get_token_from_req(req, token_location)
}
\arguments{
\item{req}{plumber request object.}

\item{token_location}{character. Location of token. Either "header" or "cookie". See details.}
}
\value{
token
}
\description{
helper function that extracts the token from the request object based
on the token_location argument.
}
\details{
Specify "header" for the \code{token_location} argument if the token
is stored in the HTTP Authorization header. Specify "cookie" if the token is stored as an
encrypted session cookie called "token" (note that unencrypted cookies are not supported).
See the \href{https://www.rplumber.io/docs/rendering-and-output.html#encrypted-cookies}{plumber docs}
for how to set an encrypted cookie.
}
\seealso{
\url{https://www.rplumber.io/docs/rendering-and-output.html#encrypted-cookies}
}
