% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth2_google.R
\name{is_authed_oauth2_google}
\alias{is_authed_oauth2_google}
\title{JWT Strategy}
\usage{
is_authed_oauth2_google(req, res, client_id, hd = NULL,
  jwks_uri = "https://www.googleapis.com/oauth2/v3/certs")
}
\arguments{
\item{req}{Request object.}

\item{res}{Response object.}

\item{client_id}{character.}

\item{hd}{character.}

\item{jwks_uri}{character.}
}
\description{
This function implements a JWT authentication strategy. The function can be used as a filter in front
of the routes. The strategy uses extracts the token from the HTTP Authorization header with the scheme 'bearer'.
}
\examples{
\dontrun{
pr$filter("sealr-jwt", function (req, res) {
  sealr::oauth2_google(req = req, res = res, secret = secret)
})
}

}
