% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_authed_return_list}
\alias{is_authed_return_list}
\title{generates a list used as return value by the is_authed_* functions}
\usage{
is_authed_return_list(is_authed, status_description = NULL,
  status_code = NULL, message = NULL)
}
\arguments{
\item{is_authed}{logical. Should be either TRUE or FALSE.}

\item{status_description}{character. Short description of the HTTP status code, e.g. 'Failed' for 401. Default NULL.}

\item{status_code}{integer. HTTP status code to return to the user. Default NULL.}

\item{message}{character. Longer description to return to the user. Default NULL.}
}
\value{
list with the following elements: is_authed, status, code, message.
}
\description{
generates a list used as return value by the is_authed_* functions
}
