% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_authed_return_list_401}
\alias{is_authed_return_list_401}
\title{Small convenience function that wraps \code{is_authed_return_list} for the
common case in this package of a "401 - Authentication required" response}
\usage{
is_authed_return_list_401()
}
\description{
Small convenience function that wraps \code{is_authed_return_list} for the
common case in this package of a "401 - Authentication required" response
}
