% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_jwt.R
\name{is_jwt_expired}
\alias{is_jwt_expired}
\title{This function checks whether a JWT is expired.}
\usage{
is_jwt_expired(payload)
}
\arguments{
\item{payload}{list. Payload of JWT.}
}
\value{
TRUE if JWT is expired, FALSE if not
(either current time < expiration time or no exp claim in JWT).
}
\description{
This function checks whether a JWT is expired.
}
