% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jwt_split}
\alias{jwt_split}
\title{adapted from https://github.com/jeroen/jose/blob/master/R/jwt.R until the function is exported in
CRAN version (version on github already exports this function)}
\usage{
jwt_split(jwt)
}
\arguments{
\item{jwt}{raw. JWT.}
}
\description{
adapted from https://github.com/jeroen/jose/blob/master/R/jwt.R until the function is exported in
CRAN version (version on github already exports this function)
}
