% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotStrentchLegend.R
\name{PlotStretchLegend}
\alias{PlotStretchLegend}
\title{plot raster images with a nice stretch and legend}
\usage{
PlotStretchLegend(r, pal, digits = 0, ...)
}
\description{
plot raster images with a nice stretch and legend
}
