% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seal.R
\name{seal}
\alias{seal}
\alias{seal}
\title{Sealing the test results}
\usage{
seal(test, load_testthat = FALSE, clip = TRUE)
}
\arguments{
\item{test}{test}

\item{load_testthat}{include \code{library(testthat)} when \emph{TRUE}}

\item{clip}{If \emph{TRUE} will overwrite the system clipboard.}
}
\description{
aaa
}
\examples{
\dontrun{
tests <- rlang::expr_interp(paste('test_that("aa", {
expect_equal(', tests, ')})', sep = "\\n"))
seal(tests, load_testthat = TRUE)
seal(tests, load_testthat = FALSE)
seal(tests, load_testthat = FALSE, clip = FALSE)
}
}
