% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothScatter_J.R
\name{smoothScatter_J}
\alias{smoothScatter_J}
\title{do smoothScatter with a linear equation legend
Usage is the same with basic smoothScatter}
\usage{
smoothScatter_J(x, y, nbin = 300, nrpoints = 0, ...)
}
\description{
do smoothScatter with a linear equation legend
Usage is the same with basic smoothScatter
}
