% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{authenticate
\code{authenticate} should be used in context of}
\usage{
authenticate(strategy, req, res, ...)
}
\arguments{
\item{strategy}{character. Must be one of: \code{jwt, google}.}

\item{req}{plumber request object}

\item{res}{plumber response object}

\item{...}{arguments to be passed down to the is_authed_ function.}
}
\description{
authenticate
\code{authenticate} should be used in context of
}
