% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_vars.R
\name{catch_vars}
\alias{catch_vars}
\alias{store_to_env}
\alias{catch_varnames}
\alias{catch_varname}
\title{catch up data.frame variable names and value}
\usage{
store_to_env(x)

catch_varnames(x)

catch_varname(x, variable)
}
\arguments{
\item{x}{\R{} object}

\item{variable}{target data.frame's column name (\code{character})}
}
\description{
catch up data.frame variable names and value
}
\examples{
\dontrun{
# Storage R object to a temporary environment
store_to_env(iris)

catch_varnames(mtcars)

# Evaluate
library(glue)
evaluate(catch_varname(mtcars, "mpg"), .GlobalEnv)
}
}
