% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jwt.R
\name{check_all_claims}
\alias{check_all_claims}
\title{This function checks that all claims passed by \code{...} to the jwt function
are correct.}
\usage{
check_all_claims(token, claims)
}
\arguments{
\item{token}{JWT extracted with jose::jwt_decode_hmac.}

\item{claims}{named list of claims to check in the JWT}
}
\value{
TRUE if the all claims are present in the JWT, FALSE if not.
}
\description{
This function checks that all claims passed by \code{...} to the jwt function
are correct.
}
