% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect}
\alias{collect}
\alias{filter_context}
\title{Collect \R{} object}
\usage{
filter_context(context = "function")
}
\arguments{
\item{context}{Types of \R{} object (\code{character})}
}
\description{
Collect \R{} object
}
\details{
\itemize{
\item filter_context
}
}
\examples{
\dontrun{
my_iris <- iris
filter_context("data.frame") \%>\%
  purrr::pmap(~ dim(..2))

my_mtcars <- mtcars
filter_context("data.frame") \%>\%
purrr::pmap(~ names(..2))
}
}
