% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-sf.R
\name{design-sf}
\alias{design-sf}
\alias{design_geom_type}
\alias{design_crs}
\alias{design_sf_dimension}
\title{Print an \R{} sf object test result}
\usage{
design_geom_type(x, ...)

design_crs(x, ...)

design_sf_dimension(x, ...)
}
\arguments{
\item{x}{\R{} object}

\item{...}{Arguments passed on to \code{seal}
\describe{
  \item{load_testthat}{include \code{library(testthat)} when \emph{TRUE}}
  \item{clip}{If \emph{TRUE} will overwrite the system clipboard.
When clipr is not available, The clip arguments is forcibly \emph{FALSE}.}
  \item{ts}{include comments that timestamp?}
}}
}
\description{
Print an \R{} sf object test result
}
\examples{
\dontrun{
library(sf)
sf_point <- st_point(c(0, 0))
sf_line <- st_linestring(matrix(1:12, ncol = 3))

design_geom_type(sf_point)
design_geom_type(sf_line)
}
\dontrun{
library(sf)
nc <-
  system.file("shape/nc.shp", package = "sf") \%>\%
  st_read()
design_crs(nc)
}
\dontrun{
library(sf)
sf_point <- st_point(c(0, 0))

design_sf_dimension(sf_point, 0)

# Failed
design_sf_dimension(sf_point, 1)
}
}
