% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{design}
\alias{design}
\alias{design_length}
\alias{design_range}
\alias{design_dim}
\alias{design_unique}
\alias{desing_missings}
\alias{design_levels}
\alias{design_nlevels}
\alias{design_nrow}
\alias{design_ncol}
\alias{design_names}
\alias{design_varclass}
\alias{design_class}
\title{Print an \R{} object test result}
\usage{
design_length(x)

design_range(x)

design_dim(x)

design_unique(x)

desing_missings(x)

design_levels(x)

design_nlevels(x)

design_nrow(x)

design_ncol(x)

design_names(x)

design_varclass(x)

design_class(x, environment = NULL)
}
\arguments{
\item{x}{\R{} object}

\item{environment}{which environment (work space) to search the available objects}
}
\description{
Print an \R{} object test result
}
