% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{design}
\alias{design}
\alias{design_expect_length}
\alias{design_expect_range}
\alias{design_expect_dim}
\alias{design_expect_unique_length}
\alias{design_expect_levels}
\alias{design_expect_nlevels}
\alias{design_expect_nrow}
\alias{design_expect_ncol}
\alias{design_expect_names}
\alias{design_expect_varclass}
\alias{design_expect_class}
\title{Print an \R{} object test result}
\usage{
design_expect_length(x)

design_expect_range(x)

design_expect_dim(x)

design_expect_unique_length(x)

design_expect_levels(x)

design_expect_nlevels(x)

design_expect_nrow(x)

design_expect_ncol(x)

design_expect_names(x)

design_expect_varclass(x)

design_expect_class(x)
}
\arguments{
\item{x}{\R{} object}
}
\description{
Print an \R{} object test result
}
