% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jwt.R
\name{jwt}
\alias{jwt}
\title{JWT Strategy}
\usage{
jwt(req, res, secret)
}
\arguments{
\item{req}{Request object.}

\item{res}{Response object.}

\item{secret}{character. This should be the secret that use to sign your JWT. The secret is converted
to raw bytes in the function.}
}
\value{

}
\description{
This function implements a JWT authentication strategy. The function can be used as a filter in front
of the routes. The strategy uses extracts the token from the HTTP Authorization header with the scheme 'bearer'.
}
