% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe.R
\name{transcribe}
\alias{transcribe}
\title{Transcribe \R{} object assert conditions}
\usage{
transcribe(x, seal = TRUE, ...)
}
\arguments{
\item{x}{\R{} object}

\item{seal}{which output testthat script}

\item{...}{Further arguments to \code{\link[=seal]{seal()}}}
}
\description{
Transcribe \R{} object assert conditions
}
\examples{
\dontrun{
x <- 1:3L; transcribe(x)
transcribe(3.14, seal = FALSE)
transcribe(letters, load_testthat = TRUE, ts = FALSE)
x <- iris$Species; transcribe(x)
transcribe(iris)
}
}
