% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe.R
\name{transcribe}
\alias{transcribe}
\title{Transcribe \R{} object assert conditions}
\usage{
transcribe(x)
}
\arguments{
\item{x}{\R{} object}
}
\description{
Transcribe \R{} object assert conditions
}
\examples{
\dontrun{
x <- 1:3L; transcribe(x) \%>\% seal()
transcribe(3.14) \%>\% seal()
transcribe(letters) \%>\% seal()
x <- iris$Species; transcribe(x) \%>\% seal()
transcribe(iris) \%>\% seal()
}
}
